+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -10,7 +10,8 @@ CHANGELOG
 2.8
 ---
 
+ * Added a $timestampAsDateTime argument to the Yaml::parse() and Yaml::dump() methods.
  * The ability to pass $timestampAsDateTime = false to the Yaml::parse method is
    deprecated since version 2.8. The argument will be removed in 3.0. Pass true instead.
- * The ability to pass $dateTimeSupport = false to the Yaml::dump method is deprecated
+ * The ability to pass $timestampAsDateTime = false to the Yaml::dump method is deprecated
    since version 2.8. The argument will be removed in 3.0. Pass true instead.